Myney/backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scipy
google-genai
python-dotenv
# Feature 1 – Unified Wallet
yfinance
curl_cffi  # required by yfinance 1.x HTTP backend
python-jose[cryptography]
passlib[bcrypt]
gunicorn